+++ svm.py
@@ -1,8 +1,18 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""
+Fake news detection
+The SVM model
+"""
+
+
 from getEmbeddings import getEmbeddings
 import numpy as np
 from sklearn.svm import SVC
 import matplotlib.pyplot as plt
 import scikitplot.plotters as skplt
+import os
 
 
 def plot_cmat(yte, ypred):
@@ -10,18 +20,23 @@ def plot_cmat(yte, ypred):
     skplt.plot_confusion_matrix(yte,ypred)
     plt.show()
 
-
-xtr,xte,ytr,yte = getEmbeddings("datasets/train.csv")
-np.save('./xtr', xtr)
-np.save('./xte', xte)
-np.save('./ytr', ytr)
-np.save('./yte', yte)
+# Read the data
+if not os.path.isfile('./xtr.npy') or \
+    not os.path.isfile('./xte.npy') or \
+    not os.path.isfile('./ytr.npy') or \
+    not os.path.isfile('./yte.npy'):
+    xtr,xte,ytr,yte = getEmbeddings("datasets/train.csv")
+    np.save('./xtr', xtr)
+    np.save('./xte', xte)
+    np.save('./ytr', ytr)
+    np.save('./yte', yte)
 
 xtr = np.load('./xtr.npy')
 xte = np.load('./xte.npy')
 ytr = np.load('./ytr.npy')
 yte = np.load('./yte.npy')
 
+# Use the built-in SVM for classification
 clf = SVC()
 clf.fit(xtr, ytr)
 y_pred = clf.predict(xte)
@@ -29,4 +44,5 @@ m = yte.shape[0]
 n = (yte != y_pred).sum()
 print("Accuracy = " + format((m-n)/m*100, '.2f') + "%")   # 88.42%
 
+# Draw the confusion matrix
 plot_cmat(yte, y_pred)
